+++ release_notes.md
@@ -12,7 +12,8 @@
 
 ### ✨ New Functionality
 
-- 
+- Support the [OAuth Refresh Token](https://help.sap.com/docs/connectivity/sap-btp-connectivity-cf/oauth-refresh-token-authentication) authentication type of the Destination Service.
+  Find more details on how to use the functionality in the [documentation](https://sap.github.io/cloud-sdk/docs/java/features/connectivity/btp-destination-service#about-the-destinationservice).
 
 ### 📈 Improvements